robot_defines formatting and style update [WIP]
- Added macros for ADC usage
- renamed some defines to new style. Macro defines use camel case as they are executing code
- renamed robotdefines.* to robot_defines.* to match current filenaming trend of underscores for spaces. File name changes for robot_defines.h carries over to many files, all of which have subsequently been updated.
