+++ RobotNoASF/communication.c
@@ -16,7 +16,7 @@
 #include <string.h>
 #include "sam.h"
 #include "communication.h"
-#include "robotdefines.h"
+#include "robot_defines.h"
 
 // XBee flow control bytes
 #define FRAME_DELIMITER 0x7E
