+++ RobotNoASF/motor_driver.h
@@ -21,7 +21,7 @@
 #define MOTOR_DRIVER_H_
 
 #include <fastmath.h>
-#include "robotdefines.h"
+#include "robot_defines.h"
 
 #define CW	0
 #define CCW 1
